feat(oauth2): Add commands for adding and deleting clients
Refactor the code for doing that from the controller to a seperate
service.
